requirements.txt
torch
transformers
datasets
scikit-learn
pandas
numpy
tqdm
pyyaml
pytorch-crf 